Add pw reset page, tooltips, confirm dialog, initial attr visualizer, QOL, grader error handling

+++ excel-autograder-frontend/src/app/views/wizard/wizard.component.ts
@@ -6,6 +6,7 @@ import {
 } from '@angular/core';
 import { ActivatedRoute, ParamMap } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
 import { AssignmentService } from '../../models/assignment/assignment.service';
 import { Assignment } from '../../models/assignment/assignment';
 import { AssignmentFactory } from '../../models/assignment/assignment.factory';
@@ -18,6 +19,9 @@ import {
 import { Question } from '../../models/question/question';
 import { HeaderComponent } from '../../models/question/facet/header/header/header.component';
 import { FacetLibrary, FacetType } from '../../models/question/facet/types/lib';
+import {
+  ConfirmationDialogComponent,
+} from '../../components/confirmation-dialog/confirmation-dialog.component';
 
 @Component({
   selector: 'app-wizard',
@@ -43,6 +47,7 @@ export class WizardComponent implements AfterViewInit {
     public questionService: QuestionService,
     public assignmentFactory: AssignmentFactory,
     private snackBar: MatSnackBar,
+    private dialog: MatDialog,
   ) { }
 
   ngAfterViewInit() {
@@ -77,9 +82,6 @@ export class WizardComponent implements AfterViewInit {
     if (!this.activeAssignment) return;
     this.activeAssignment.questions = this.questionService.getQuestions();
     this.activeAssignment.save().subscribe({
-      next: () => {
-        this.snackBar.open('Saved!', 'Close', { duration: 1500 });
-      },
       error: () => {
         this.snackBar.open('Failed to save!', 'Close', { duration: 1500 });
       },
@@ -87,13 +89,25 @@ export class WizardComponent implements AfterViewInit {
   }
 
   deleteActiveQuestion() {
-    if (!this.activeQuestion) return;
-    const question = this.activeQuestion;
-    if (this.questionService.getQuestions().length === 1) this.setActiveQuestion(null);
-    let newIdx = this.questionService.getQuestions().indexOf(question) - 1;
-    if (newIdx < 0) newIdx = 0;
-    this.questionService.removeQuestion(question);
-    this.setActiveQuestion(this.questionService.getQuestions()[newIdx]);
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      width: '350px',
+      data: {
+        title: 'Confirm Delete',
+        message: 'Any data on this question will be lost',
+      },
+    });
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        if (!this.activeQuestion) return;
+        const question = this.activeQuestion;
+        if (this.questionService.getQuestions().length === 1) this.setActiveQuestion(null);
+        let newIdx = this.questionService.getQuestions().indexOf(question) - 1;
+        if (newIdx < 0) newIdx = 0;
+        this.questionService.removeQuestion(question);
+        this.setActiveQuestion(this.questionService.getQuestions()[newIdx]);
+        this.saveQuestions();
+      }
+    });
   }
 
   addQuestion() {
@@ -101,6 +115,7 @@ export class WizardComponent implements AfterViewInit {
     this.setActiveQuestion(
       this.questionService.getQuestions()[this.questionService.getQuestions().length - 1],
     );
+    this.saveQuestions();
   }
 
   prevQuestion() {
@@ -135,11 +150,13 @@ export class WizardComponent implements AfterViewInit {
     const header = this.facetContainer.createComponent(HeaderComponent);
     const com = FacetLibrary.getFacetComponent(facet, this.facetContainer);
     com.setInput('facet', facet);
+    com.instance.valueChange.subscribe(() => { this.saveQuestions(); });
     com.setInput('workbookService', this.workbookService);
     header.setInput('question', this.activeQuestion);
     header.setInput('facet', facet);
     header.setInput('component', com);
     header.setInput('self', header);
+    this.saveQuestions();
   }
 
   protected readonly FacetType = FacetType;
